requirements.txt
rich
python-vlc